+++ ServiceDesk/ServiceDesk/App_Code/db.cs
@@ -62,17 +62,5 @@ namespace ServiceDesk
             conn.Open();
             cmd.ExecuteNonQuery();
         }
-
-        public static void CreateUser(string data)
-        {
-            string SQLComm = "Insert into bruger (fornavn,efternavn,password,telefon,type,email) VALUES (" + data + ")";
-            db.SQLNonQuery(SQLComm);
-        }
-
-        public static void Deleteuser(int id)
-        {
-            string SQLComm = "Delete from brugere where id=" + id;
-            db.SQLNonQuery(SQLComm);
-        }
     }
 }
